Redesigned landing page
I changed the background to be a single shade of purple and changed the title/subtitle colors. I also added social media icons which will soon have links. I need to finish cleaning up the roadmap then I will work on fonts and text. We also need the summary and roadmap to appear as the user scrolls the page.